+++ ricaun.Revit.Installation.Tests/ApplicationPluginsUtils_Tests.cs
@@ -1,8 +1,11 @@
 ﻿using NUnit.Framework;
+using NUnit.Framework.Constraints;
+using NUnit.Framework.Internal;
 using ricaun.Revit.Installation.Tests.Utils;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -10,12 +13,25 @@ namespace ricaun.Revit.Installation.Tests
 {
     public class ApplicationPluginsUtils_Tests
     {
+        [OneTimeSetUp]
+        public void Setup()
+        {
+            // Create a temporary folder for the tests
+            applicationPluginsFolder = Path.Combine(Path.GetTempPath(), $"ApplicationPlugins_{Guid.NewGuid().ToString("N")}");
+            if (Directory.Exists(applicationPluginsFolder))
+            {
+                Directory.Delete(applicationPluginsFolder, true);
+            }
+            Directory.CreateDirectory(applicationPluginsFolder);
+        }
+
+        private string applicationPluginsFolder;
+
         [TestCase("RevitAddin.DA.Tester")]
         public async Task ApplicationPluginsUtils_Test_Download_Async(string projectName)
         {
             var bundleUrl = $@"https://github.com/ricaun-io/{projectName}/releases/latest/download/{projectName}.bundle.zip";
 
-            var applicationPluginsFolder = Path.Combine(Path.GetTempPath(), "ApplicationPlugins");
             var bundleName = Path.GetFileNameWithoutExtension(bundleUrl);
 
             Console.WriteLine($"DownloadBundle: {bundleName}");
@@ -45,7 +61,8 @@ namespace ricaun.Revit.Installation.Tests
         {
             var bundleUrl = $@"https://github.com/ricaun-io/{projectName}/releases/latest/download/{projectName}.bundle.zip";
 
-            var applicationPluginsFolder = Path.Combine(Path.GetTempPath(), "ApplicationPlugins");
+            //applicationPluginsFolder = RevitUtils.GetCurrentUserApplicationPluginsFolder();
+
             var bundleName = Path.GetFileNameWithoutExtension(bundleUrl);
 
             Console.WriteLine($"DownloadBundle: {bundleName}");
@@ -54,17 +71,15 @@ namespace ricaun.Revit.Installation.Tests
             {
                 Console.WriteLine(e);
                 Assert.Fail(e.Message);
-            }, (log) =>
-            {
-                Console.WriteLine(log);
             });
+            //}, Log);
 
             Console.WriteLine($"Bundle Exists: {Directory.Exists(Path.Combine(applicationPluginsFolder, bundleName))}");
             Assert.IsTrue(Directory.Exists(Path.Combine(applicationPluginsFolder, bundleName)));
 
             Thread.Sleep(1000);
 
-            ApplicationPluginsUtils.DeleteBundle(applicationPluginsFolder, bundleName);
+            ApplicationPluginsUtils.DeleteBundle(applicationPluginsFolder, bundleName, Log);
             Console.WriteLine($"Bundle Exists: {Directory.Exists(Path.Combine(applicationPluginsFolder, bundleName))}");
         }
 
@@ -74,9 +89,9 @@ namespace ricaun.Revit.Installation.Tests
         [TestCase("FakeBundle", true, false)]
         public void ApplicationPluginsUtils_Test_BundleCreatorUtils(string projectName, bool includeBundleDirectory, bool includeContents = true)
         {
+            projectName += Guid.NewGuid().ToString("N");
             var bundleUrl = BundleCreatorUtils.CreateBundleZip(projectName, includeBundleDirectory, includeContents);
 
-            var applicationPluginsFolder = Path.Combine(Path.GetTempPath(), "ApplicationPlugins");
             var bundleName = Path.GetFileNameWithoutExtension(bundleUrl);
 
             Console.WriteLine($"DownloadBundle: {bundleName}");
@@ -90,13 +105,80 @@ namespace ricaun.Revit.Installation.Tests
                 Console.WriteLine(log);
             });
 
-            Console.WriteLine($"Bundle Exists: {Directory.Exists(Path.Combine(applicationPluginsFolder, bundleName))}");
-            Assert.IsTrue(Directory.Exists(Path.Combine(applicationPluginsFolder, bundleName)));
+            var bundleDirectory = Path.Combine(applicationPluginsFolder, bundleName);
+            var bundlePackageContentsPath = Path.Combine(bundleDirectory, "PackageContents.xml");
+
+            Console.WriteLine($"Bundle Exists: {Directory.Exists(bundleDirectory)}");
+            Assert.IsTrue(Directory.Exists(bundleDirectory));
+            Assert.IsTrue(File.Exists(bundlePackageContentsPath));
 
             Thread.Sleep(1000);
 
             ApplicationPluginsUtils.DeleteBundle(applicationPluginsFolder, bundleName);
-            Console.WriteLine($"Bundle Exists: {Directory.Exists(Path.Combine(applicationPluginsFolder, bundleName))}");
+            Console.WriteLine($"Bundle Exists: {Directory.Exists(bundleDirectory)}");
+
+            Assert.IsFalse(Directory.Exists(bundleDirectory));
         }
+
+        [TestCase("FakeBundleUsedByProcess", 0)]
+        [TestCase("FakeBundleUsedByProcess", 1)]
+        [TestCase("FakeBundleUsedByProcess", 2)]
+        public void ApplicationPluginsUtils_Test_BundleCreatorUtils_FileUsedByProcess(string projectName, int numberFile)
+        {
+            projectName += Guid.NewGuid().ToString("N");
+            var bundleUrl = BundleCreatorUtils.CreateBundleZip(projectName);
+
+            var bundleName = Path.GetFileNameWithoutExtension(bundleUrl);
+
+            Console.WriteLine($"DownloadBundle: {bundleName}");
+
+            ApplicationPluginsUtils.DownloadBundle(applicationPluginsFolder, bundleUrl, (e) =>
+            {
+                Console.WriteLine(e);
+                Assert.Fail(e.Message);
+            }, Log);
+
+            var bundleDirectory = Path.Combine(applicationPluginsFolder, bundleName);
+            var bundlePackageContentsPath = Path.Combine(bundleDirectory, "PackageContents.xml");
+
+            Console.WriteLine($"Bundle Exists: {Directory.Exists(bundleDirectory)}");
+            Assert.IsTrue(Directory.Exists(bundleDirectory));
+            Assert.IsTrue(File.Exists(bundlePackageContentsPath));
+
+            Thread.Sleep(1000);
+
+            var fileName = "File.xml";
+            var fileNumberName = $"Number_{numberFile}.xml";
+            var files = Directory.GetFiles(bundleDirectory, fileNumberName, SearchOption.AllDirectories);
+            Assert.IsTrue(files.Length > 0, $"File '{fileNumberName}' not found in bundle directory.");
+            var streams = files.Select(file => File.Open(file, FileMode.Open, FileAccess.Read, FileShare.None)).ToArray();
+            var filesCount = files.Select(file => Directory.GetFiles(Path.GetDirectoryName(file), "*").Length).Sum();
+            
+            ApplicationPluginsUtils.DeleteBundle(applicationPluginsFolder, bundleName, Log);
+            Console.WriteLine($"Bundle Exists: {Directory.Exists(bundleDirectory)}");
+
+            Assert.IsTrue(Directory.Exists(bundleDirectory));
+            Assert.IsFalse(File.Exists(bundlePackageContentsPath));
+
+            var filesCountAfter = files.Select(file => Directory.GetFiles(Path.GetDirectoryName(file), "*").Length).Sum();
+            Assert.AreEqual(filesCount, filesCountAfter, "Files were deleted while they were still in use.");
+            foreach (var file in files)
+            {
+                var fileNamePath = Path.Combine(Path.GetDirectoryName(file), fileName);
+                Assert.IsTrue(File.Exists(fileNamePath), "File was deleted while is not been used, but some file and the same folder is been used.");
+            }
+
+            foreach (var stream in streams)
+            {
+                stream.Dispose();
+            }
+
+            ApplicationPluginsUtils.DeleteBundle(applicationPluginsFolder, bundleName, Log);
+            Console.WriteLine($"Bundle Exists: {Directory.Exists(bundleDirectory)}");
+
+            Assert.IsFalse(Directory.Exists(bundleDirectory));
+        }
+
+        private static void Log(string message) => Console.WriteLine(message);
     }
 }